Fix out by one error and add new helper method for getting entities form other scuttlebutt instances
