+++ src/test/scala/org/openlaw/scuttlebutt/persistence/Types.scala
@@ -1,0 +1,11 @@
+package org.openlaw.scuttlebutt.persistence
+
+case class Cmd(data: String)
+
+case class Evt(data: String)
+
+case class ExampleState(events: List[String] = Nil) {
+  def updated(evt: Evt): ExampleState = copy(evt.data :: events)
+  def size: Int = events.length
+  override def toString: String = events.reverse.toString
+}
